+++ khal/ui/colors.py
@@ -19,17 +19,18 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+from typing import Dict, List, Tuple
 
 dark = [
     ('header', 'white', 'black'),
     ('footer', 'white', 'black'),
     ('line header', 'black', 'white', 'bold'),
     ('alt header', 'white', '', 'bold'),
-    ('bright', 'dark blue', 'white', ('bold', 'standout')),
+    ('bright', 'dark blue', 'white', 'bold,standout'),
     ('list', 'black', 'white'),
     ('list focused', 'white', 'light blue', 'bold'),
     ('edit', 'black', 'white'),
-    ('edit focused', 'white', 'light blue', 'bold'),
+    ('edit focus', 'white', 'light blue', 'bold'),
     ('button', 'black', 'dark cyan'),
     ('button focused', 'white', 'light blue', 'bold'),
 
@@ -44,13 +45,17 @@ dark = [
     ('dayname', 'light gray', ''),
     ('monthname', 'light gray', ''),
     ('weeknumber_right', 'light gray', ''),
-    ('edit', 'white', 'dark blue'),
     ('alert', 'white', 'dark red'),
     ('mark', 'white', 'dark green'),
     ('frame', 'white', 'black'),
     ('frame focus', 'light red', 'black'),
     ('frame focus color', 'dark blue', 'black'),
     ('frame focus top', 'dark magenta', 'black'),
+
+    ('eventcolumn', '', '', ''),
+    ('eventcolumn focus', '', '', ''),
+    ('calendar', '', '', ''),
+    ('calendar focus', '', '', ''),
 
     ('editbx', 'light gray', 'dark blue'),
     ('editcp', 'black', 'light gray', 'standout'),
@@ -63,11 +68,11 @@ light = [
     ('footer', 'black', 'white'),
     ('line header', 'black', 'white', 'bold'),
     ('alt header', 'black', '', 'bold'),
-    ('bright', 'dark blue', 'white', ('bold', 'standout')),
+    ('bright', 'dark blue', 'white', 'bold,standout'),
     ('list', 'black', 'white'),
     ('list focused', 'white', 'light blue', 'bold'),
     ('edit', 'black', 'white'),
-    ('edit focused', 'white', 'light blue', 'bold'),
+    ('edit focus', 'white', 'light blue', 'bold'),
     ('button', 'black', 'dark cyan'),
     ('button focused', 'white', 'light blue', 'bold'),
 
@@ -76,13 +81,12 @@ light = [
     ('today', 'black', 'light gray'),
 
     ('date header', '', 'white'),
-    ('date header focused', 'white', 'dark gray', ('bold', 'standout')),
+    ('date header focused', 'white', 'dark gray', 'bold,standout'),
     ('date header selected', 'dark gray', 'light cyan'),
 
     ('dayname', 'dark gray', 'white'),
     ('monthname', 'dark gray', 'white'),
     ('weeknumber_right', 'dark gray', 'white'),
-    ('edit', 'white', 'dark blue'),
     ('alert', 'white', 'dark red'),
     ('mark', 'white', 'dark green'),
     ('frame', 'dark gray', 'white'),
@@ -90,9 +94,16 @@ light = [
     ('frame focus color', 'dark blue', 'white'),
     ('frame focus top', 'dark magenta', 'white'),
 
+    ('eventcolumn', '', '', ''),
+    ('eventcolumn focus', '', '', ''),
+    ('calendar', '', '', ''),
+    ('calendar focus', '', '', ''),
+
     ('editbx', 'light gray', 'dark blue'),
     ('editcp', 'black', 'light gray', 'standout'),
     ('popupbg', 'white', 'black', 'bold'),
     ('popupper', 'black', 'light gray'),
     ('caption', 'black', '', ''),
 ]
+
+themes: Dict[str, List[Tuple[str, ...]]] = {'light': light, 'dark': dark}